Clarify the p-card limits of training spending
Trainings over $3500 trigger a full procurement process and it'd be good to point that out before people get too far along in the process. We might also consider telling people they can mitigate that by agreeing to pay the excess balance. This is helpful if the training is, say, $3501. Someone might choose to pay the extra dollar out of pocket rather than do market research the full gamut of procurement. We'd have to word it carefully to make it clear this is simply *an* option, not a requirement for more costly trainings.