consolidate information about software requests
There are explanations all over the Handbook about how to request
various software, which are hard to keep in sync. This points them all
to a single section, which points to a single doc. Note that this latter
part is temporary - at least part of it will be moved into the Handbook
soon, once it's more polished.