Adding  `vpn.gsa.gov/adcontractors` vpn url
The previous VPN url "vpn.gsa.gov/gfeotp" did not work for me ever, so I couldn't use Anyconnect, and Cristina turned me on to "vpn.gsa.gov/adcontractors" which works and makes this a functional system! I don't know if the old one is completely depredated or where it originally came from, so I'm keeping it in there but noting it's older and may not work.