Update travel-guide-1-book-travel.md
It may just be me but I can't get the links to work. 

I've put forth some thoughts to shorten grammar 

I think there are areas where the documents structure could be improved.  In areas it isn't consistent.  I've tried to make it more consistent throughout.  But I could give it another pass.  

We may want to consider marking up the images with arrows or circles to highlight relevant information.  this may just be because I'm not in Concur as I'm going through the instructions though.