Notes
Maybe add a more prominent Title to this doc "Things to Consider while Traveling"  It took me a minute to understand that this is want this doc is about.  I wonder if this would be easier to read as a "Dos and Don'ts" document.  Ex. Do use your travel card, don't forget to do get your voucher.