Update travel-guide-2-secure-approvals.md
Should the email templates go in a separate document or documents?  It seems like the middle of this handbook page is an odd place to find them.  Should we pull questions from the different parts of this doc and put them together in a general FAQ section.  There are so many questions that its a little distracting from the larger process.