Update training-and-development.md
Removed the free training section and instead updated the first section to ## Free training and development opportunities where it encompasses all of the free trainings (including coaching/mentoring). Added another supporting sentence to the mentor program section. Updated the main headline to be more general to apply to all of the content on the page (which also follows the pattern of most other Handbook pages).