Add OTP generation to setup
While setting up my own VPN, I got tripped up by not having the one time password client installed on my phone.  I thought the OTP that it was asking for was one of the many other passwords asked for in the setup process.  Figured it would help others doing this setup to spell out the need for the phone app here.