Create "Guidance on hosting non-government speakers" page
TTS teams, guilds, and working groups often bring in external speakers but guidance is buried in emails and provided by a colleague who has termed-out and is unavailable for guidance. Outreach team unofficially offered guidance on this process, but the approvals team was unsure of (background conversation in Slack here: https://gsa-tts.slack.com/archives/C02APU5RT/p1571415750010100). I'd love to add this page under "18F team resources" so that external speakers are brought in ethically and compliantly.