Update tock.md
Making an update to the weekly billable hour expectation section on 18F non-billable work. It may not be as clear as it can be for folks who attend professional development or other training courses that they use 18F non-billable. There have been a few instances in the past weeks where folks are tocking to "TTS OA / Training and Conferences" for conference and training events.