Changed process for software authorization
This PR is very limited scoped and focused on the 3 step process. 

Putting together some information for the SaaS authorization infographic and realized that a small tweak to the process can go a long way and would be less of a deterrent for folks to enter into the process. Also, it'll be one less place to change the General Request form when we do that pull request. These changes are in line with the process for buying technology that are over $10k as well as the process that the VA doe sit in their [VA SaaS Playbook](https://www.oit.va.gov/services/cloud-software/saas/playbook#security)