This is more of a change of tone. (#646)
Accessibility is a process and this is about making sure we're working with folks who are already working on accessibility as a default way of our team working. We want to stand on the shoulders of giants.

[//]: # (rtdbot-start)

URL of RTD document: https://civicactions-handbook.readthedocs.io/en/mgifford-patch-2/ ![Documentation Status](https://readthedocs.org/projects/civicactions-handbook/badge/?version=mgifford-patch-2)

[//]: # (rtdbot-end)

Co-authored-by: Restyled.io <commits@restyled.io>