Change "Before you ship" link to navigate to guide homepage
Prior to these changes, the "Before you ship" link shown on the handbook index would navigate directly to the "Amazon Web Services" subpage of the Before You Ship guide. These changes propose to link to the root page of the Before You Ship guide, as this is where one might expect to be directed to learn all there is about launching software.