Redirect to participant agreement in UX guide (#3432)
The Methods site will soon be deprecating the url that currently hosts the participant agreement example and redirecting to where this example also lives in the 18F UX guide. This PR replaces the link to the example in the Methods with a link to the where the example lives in the UX guide.