update printer instructions link
The old link is dead and the [insite link](https://insite.gsa.gov/employee-resources/information-technology/hardware-equipment/connect-to-network-printers?term=printer) makes it sound like you have to make an appointment, but you can map your printer yourself and there are official instructions, that is what I linked to. The screen shots are old, but the idea is the same.