Direct link to Login help center
Users are periodically opening up handbook issues seeking Login.gov support.
This puts a direct link in the template to try to help get them back on track.

I did not make this a Markdown link because I expect it will be seen when
they are trying to open a GitHub issue in the edit view.