Update first-time-travel-complete-concur-profile.md
Concur's website has updated (the screenshots referenced in "Having Trouble?" are no longer accurate) and I have updated step 2 to reflect the current system. Also added a period in step 2 and 3 to maintain consistency with step 1, 4, and 5.

Also, apologies if this is a duplicate request, I am new to Github and didn't do a pull request so this is my second attempt.